move plugin to subfolder
